support 2.0 lr_scheduler for 2.0 optimizer
