+++ python/paddle/optimizer/lr_scheduler.py
@@ -53,8 +53,9 @@ class _LRScheduler(object):
 
     def step(self, epoch=None):
         """
-        'step' should be called after 'minimize' . It will update the learning rate in optimizer according to 'epoch'.  
-        The new learning rate will take effect on next epoch.
+        ``scheduler.step()`` should be called after ``optimizer.step()`` . 
+        It will update the learning rate in optimizer according to ``epoch`` .  
+        The new learning rate will take effect on next ``optimizer.step()`` .
 
         Args:
             epoch (int, None): specify current epoch. Default: None. Auto-increment from last_epoch=-1.
@@ -109,7 +110,7 @@ class _LRScheduler(object):
         """
         self.keys = ['last_epoch', 'last_lr']
 
-    def set_dict(self, state_dict):
+    def set_state_dict(self, state_dict):
         """
         Loads the schedulers state.
         """
@@ -125,9 +126,6 @@ class _LRScheduler(object):
             warnings.warn(
                 "There are some unused values in state_dict. Maybe the optimizer have different 'LearningRateDecay' when invoking state_dict and set_dict"
             )
-
-    # alias for set_dict
-    set_state_dict = set_dict
 
     def get_lr(self):
         # calculate by python float
@@ -176,8 +174,8 @@ class NoamLR(_LRScheduler):
                     out = linear(x)
                     loss = paddle.reduce_mean(out)
                     loss.backward()
-                    sgd.minimize(loss)
-                    linear.clear_gradients()
+                    sgd.step()
+                    sgd.clear_grad()
                 scheduler.step()
 
             # train on static mode
@@ -274,8 +272,8 @@ class PiecewiseLR(_LRScheduler):
                     out = linear(x)
                     loss = paddle.reduce_mean(out)
                     loss.backward()
-                    sgd.minimize(loss)
-                    linear.clear_gradients()
+                    sgd.step()
+                    sgd.clear_grad()
                 scheduler.step()
 
             # train on static mode
@@ -358,8 +356,8 @@ class NaturalExpLR(_LRScheduler):
                     out = linear(x)
                     loss = paddle.reduce_mean(out)
                     loss.backward()
-                    sgd.minimize(loss)
-                    linear.clear_gradients()
+                    sgd.step()
+                    sgd.clear_grad()
                 scheduler.step()
 
             # train on static mode
@@ -437,8 +435,8 @@ class InverseTimeLR(_LRScheduler):
                     out = linear(x)
                     loss = paddle.reduce_mean(out)
                     loss.backward()
-                    sgd.minimize(loss)
-                    linear.clear_gradients()
+                    sgd.step()
+                    sgd.clear_grad()
                 scheduler.step()
 
             # train on static mode
@@ -533,8 +531,8 @@ class PolynomialLR(_LRScheduler):
                     out = linear(x)
                     loss = paddle.reduce_mean(out)
                     loss.backward()
-                    sgd.minimize(loss)
-                    linear.clear_gradients()
+                    sgd.step()
+                    sgd.clear_grad()
                 scheduler.step()
 
             # train on static mode
@@ -649,8 +647,8 @@ class LinearLrWarmup(_LRScheduler):
                     out = linear(x)
                     loss = paddle.reduce_mean(out)
                     loss.backward()
-                    sgd.minimize(loss)
-                    linear.clear_gradients()
+                    sgd.step()
+                    sgd.clear_grad()
                 scheduler.step()
 
             # train on static mode
@@ -754,8 +752,8 @@ class ExponentialLR(_LRScheduler):
                     out = linear(x)
                     loss = paddle.reduce_mean(out)
                     loss.backward()
-                    sgd.minimize(loss)
-                    linear.clear_gradients()
+                    sgd.step()
+                    sgd.clear_grad()
                 scheduler.step()
 
             # train on static mode
@@ -842,8 +840,8 @@ class MultiStepLR(_LRScheduler):
                     out = linear(x)
                     loss = paddle.reduce_mean(out)
                     loss.backward()
-                    sgd.minimize(loss)
-                    linear.clear_gradients()
+                    sgd.step()
+                    sgd.clear_grad()
                 scheduler.step()
 
             # train on static mode
@@ -951,8 +949,8 @@ class StepLR(_LRScheduler):
                     out = linear(x)
                     loss = paddle.reduce_mean(out)
                     loss.backward()
-                    sgd.minimize(loss)
-                    linear.clear_gradients()
+                    sgd.step()
+                    sgd.clear_grad()
                 scheduler.step()
 
             # train on static mode
@@ -1047,8 +1045,8 @@ class LambdaLR(_LRScheduler):
                     out = linear(x)
                     loss = paddle.reduce_mean(out)
                     loss.backward()
-                    sgd.minimize(loss)
-                    linear.clear_gradients()
+                    sgd.step()
+                    sgd.clear_grad()
                 scheduler.step()
 
             # train on static mode
@@ -1147,8 +1145,8 @@ class ReduceLROnPlateau(_LRScheduler):
                     out = linear(x)
                     loss = paddle.reduce_mean(out)
                     loss.backward()
-                    sgd.minimize(loss)
-                    linear.clear_gradients()
+                    sgd.step()
+                    sgd.clear_grad()
                 scheduler.step(loss)
 
             # train on static mode
@@ -1238,8 +1236,9 @@ class ReduceLROnPlateau(_LRScheduler):
 
     def step(self, metrics, epoch=None):
         """
-        step should be called after 'minimize' . It will update the learning rate in optimizer according to ``metrics`` .  
-        The new learning rate will take effect on next epoch.
+        ``scheduler.step()`` should be called after ``optimizer.step()`` . 
+        It will update the learning rate in optimizer according to ``metrics`` .  
+        The new learning rate will take effect on next ``optimizer.step()`` .
 
         Args:
             metrics (Tensor|numpy.ndarray|float): Which will be monitored to determine whether the learning rate will reduce. 
@@ -1364,8 +1363,8 @@ class CosineAnnealingLR(_LRScheduler):
                     out = linear(x)
                     loss = paddle.reduce_mean(out)
                     loss.backward()
-                    sgd.minimize(loss)
-                    linear.clear_gradients()
+                    sgd.step()
+                    sgd.clear_grad()
                 scheduler.step()
 
             # train on static mode
